fix(utils): move some funcs to __init__.py
