+++ tests/test_cdis.py
@@ -1,5 +1,3 @@
-import urllib
-
 from .utils import cdis as utils
