+++ tests/test_public_keys_endpoint.py
@@ -1,7 +1,7 @@
 from collections import OrderedDict
-
-from .utils import oauth2 as utils
 from .test_settings import JWT_KEYPAIR_FILES
+
+import utils as utils
 
 
 def test_keys_endpoint(app, client):
